Add modules in structure of js files
